Walking Animation working with Asterdimon
I used the moveSpeed float variable in the move.cs in my player.cs  to
determine if Asterdimon was walking.
I basically made the walk animation play if moveSpeed was greater than
zero and transition to idle if it is less than 1. It works so far but
there are other issues with Asterdimon slowly rotating that cause the
animation transition to be buggy.